requirements.txt
anyio==4.6.0
APScheduler==3.10.4
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.2
Flask==2.3.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.0.2
pandas==2.2.3
pybit==5.8.0
pycryptodome==3.20.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-telegram-bot==21.6
pytz==2024.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
websocket-client==1.8.0
Werkzeug==3.0.4
zipp==3.20.2
